requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dicttoxml==1.7.4
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
json2xml==3.6.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
xmltodict==0.11.0
